Backend/requirements.txt

openai==1.74.0
openai-agents==0.0.11
python-dotenv==1.0.1
fastapi==0.115.12
pydantic==2.10.7
